requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.31.1
pillow==10.4.0
numpy==2.1.2
python-multipart==0.0.9
starlette==0.41.2


